[WIP] Several new features

+++ spiketoolkit/sorters/tests/test_launcher.py
@@ -22,7 +22,7 @@ def test_run_sorters_with_list():
     
     
 
-def test_run_sorters():
+def test_run_sorters_with_dict():
     rec0, _ = se.example_datasets.toy_example(num_channels=4, duration=30)
     rec1, _ = se.example_datasets.toy_example(num_channels=8, duration=30)
     
@@ -32,7 +32,7 @@ def test_run_sorters():
     #~ sorter_list = ['tridesclous',  'klusta',]
     sorter_list = ['tridesclous', ]
     
-    working_folder = 'test_run_sorters'
+    working_folder = 'test_run_sorters_dict'
     if os.path.exists(working_folder):
         shutil.rmtree(working_folder)
     
@@ -42,12 +42,26 @@ def test_run_sorters():
     run_sorters(sorter_list, recording_dict, working_folder, engine=None)
     t1 = time.perf_counter()
     print(t1-t0)
+
+
+def test_run_sorters_multiprocessing():
     
-    shutil.rmtree(working_folder)
+    recording_dict = {}
+    for i in range(8):
+        rec, _ = se.example_datasets.toy_example(num_channels=8, duration=30)
+        recording_dict['rec_'+str(i)] = rec
     
+    # sorter_list = ['mountainsort4', 'klusta', 'tridesclous']
+    #~ sorter_list = ['tridesclous',  'klusta',]
+    sorter_list = ['tridesclous', ]
+    
+    working_folder = 'test_run_sorters_mp'
+    if os.path.exists(working_folder):
+        shutil.rmtree(working_folder)
+
     # multiprocessing
     t0 = time.perf_counter()
-    run_sorters(sorter_list, recording_dict, working_folder, engine='multiprocessing')
+    run_sorters(sorter_list, recording_dict, working_folder, engine='multiprocessing', processes=4)
     t1 = time.perf_counter()
     print(t1-t0)
     
@@ -57,4 +71,6 @@ def test_run_sorters():
 if __name__ == '__main__':
     #~ test_run_sorters_with_list()
     
-    test_run_sorters()
+    #~ test_run_sorters_with_dict()
+    
+    test_run_sorters_multiprocessing()
